% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{build_delete_id_body}
\alias{build_delete_id_body}
\title{Build Request Body for Delete by Id}
\usage{
build_delete_id_body(id, entity_name)
}
\arguments{
\item{id}{character; a MS Dynamics CRM generated id}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}
}
\value{
\code{XMLNode} to be used as the body for the request
}
\description{
This function builds the XML body for a request to delete an entity by Id.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
