% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{build_retrieve_id_body}
\alias{build_retrieve_id_body}
\title{Build Request Body for Retrieve by Id}
\usage{
build_retrieve_id_body(id, entity_name, columns, all_columns)
}
\arguments{
\item{id}{character; a MS Dynamics CRM generated id}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{columns}{character; one or more strings indicating columns}

\item{all_columns}{logical; an indicator if all possible columns should be returned 
for the entity. If \code{TRUE} this parameter will override the \code{columns} parameter.}
}
\value{
\code{XMLNode} to be used as the body for the request
}
\description{
This function builds the XML body for a request to retrieve an entity by Id.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
