% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{catch_errors_wo_retry}
\alias{catch_errors_wo_retry}
\title{Another function to catch and print HTTP errors}
\usage{
catch_errors_wo_retry(x, verbose = FALSE)
}
\description{
Another function to catch and print HTTP errors
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
