% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{convert_entity_data_to_body}
\alias{convert_entity_data_to_body}
\title{Build Request Body for a Single Record}
\usage{
convert_entity_data_to_body(input_data, entity_name, operation)
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or
\code{tbl_df}; data can be coerced into a \code{data.frame}}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{operation}{character; a string defining the type of operation being 
performed (e.g. "create", "update", "delete", "associate")}
}
\value{
\code{XMLNode} to be used as the body for the request
}
\description{
This function builds the XML body for a request to create or update a single record in 
the specified entity.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
