% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dyn_auth_refresh}
\alias{dyn_auth_refresh}
\title{Refresh an existing authorized MS Dynamics CRM session}
\usage{
dyn_auth_refresh(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\description{
Force the current state to refresh. This is only needed for times when the token 
has expired
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
