% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dyn_create}
\alias{dyn_create}
\title{Create Records}
\usage{
dyn_create(input_data, entity_name, verbose = FALSE)
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or
\code{tbl_df}; data can be coerced into a \code{data.frame}}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of records with success indicator
}
\description{
Adds one or more new records to your organization’s data.
}
\examples{
\dontrun{
n <- 2
new_contacts <- tibble(firstname = rep("Test", n),
                       lastname = paste0("Contact", 1:n))
new_contacts_result <- dyn_create(new_contacts, entity_name="contact")
}
}
