% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{dyn_delete}
\alias{dyn_delete}
\title{Delete Records}
\usage{
dyn_delete(ids, entity_name, verbose = FALSE)
}
\arguments{
\item{ids}{\code{vector}, \code{matrix}, \code{data.frame}, or
\code{tbl_df}; if not a vector, there must be a column called Id (case-insensitive)
that can be passed in the request}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of records with success indicator
}
\description{
Deletes one or more records to your organization’s data.
}
\examples{
\dontrun{
n <- 2
new_contacts <- tibble(firstname = rep("Test", n),
                       lastname = paste0("Contact", 1:n))
new_contacts_result1 <- dyn_create(new_contacts, entity_name="contact")
deleted_contacts_result1 <- dyn_delete(new_contacts_result1$id,
                                       entity_name="contact")
}
}
