% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dyn_header}
\alias{dyn_header}
\title{Return authentication header}
\usage{
dyn_header(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
character; a string of the header element of the current state; otherwise NULL
}
\description{
Return authentication header
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
