% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dyn_input_data_validation}
\alias{dyn_input_data_validation}
\title{Validate the input for an operation}
\usage{
dyn_input_data_validation(input_data, operation = "")
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or
\code{tbl_df}; data can be coerced into a \code{data.frame}}

\item{operation}{character; a string defining the type of operation being 
performed (e.g. "create", "update", "delete", "associate")}
}
\description{
Validate the input for an operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
