% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{dyn_query}
\alias{dyn_query}
\title{Perform FetchXML Query}
\usage{
dyn_query(fetchxml = NULL, entity_name = NULL, attributes = NULL,
  all_attributes = FALSE, page_size = 100, top = NULL, verbose = FALSE)
}
\arguments{
\item{fetchxml}{character; a string defining a FetchXML query}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{attributes}{character; one or more strings indicating attributes}

\item{all_attributes}{logical; an indicator if all possible attributes should be returned 
for the entity. If \code{TRUE} this parameter will override the \code{attributes} parameter.}

\item{page_size}{numeric; a number indicating the records per page that are 
returned. This is provided for performance improvements for queries returning a 
large number of records.}

\item{top}{numeric; a number indicating the first N number of records that should 
be returned by the query. This behavior is similar to the LIMIT or TOP keyword in SQL.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of records
}
\description{
Executes a query against the specified object and returns data that matches
the specified criteria.
}
\note{
If the \code{fetchxml} argument is specified, then all other fetch arguments will 
be ignored. They are only provided as a convenience in case the user does not want 
to specify exact FetchXML for simple "SELECT" queries (e.g. SELECT * FROM ACCOUNT).
}
\examples{
\dontrun{
fetchxml <- "
<fetch version='1.0' count='100' page='1'>
  <entity name='systemuser'>
    <attribute name='fullname'/>
    <attribute name='internalemailaddress'/>
    <attribute name='systemuserid'/>
    <order attribute='fullname' descending='false'/>
  </entity>
</fetch>"
users <- dyn_query(fetchxml)

# using arguments instead of explicit FetchXML statement
users <- dyn_query(entity_name="systemuser",
                   attributes=c("fullname", "internalemailaddress"))
}
}
