% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{dyn_update}
\alias{dyn_update}
\title{Update Records}
\usage{
dyn_update(input_data, entity_name, verbose = FALSE)
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or
\code{tbl_df}; data can be coerced into a \code{data.frame}}

\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tbl_df} of records with success indicator
}
\description{
Updates one or more records to your organization’s data.
}
\examples{
\dontrun{
n <- 2
new_contacts <- tibble(firstname = rep("Test", n),
                       lastname = paste0("Contact", 1:n))
new_contacts_result <- dyn_create(new_contacts, entity_name="contact")

update_contacts <- tibble(firstname = rep("TestTest", n),
                          lastname = paste0("Contact", 1:n),
                          id = new_contacts_result$id)
updated_contacts_result1 <- dyn_update(update_contacts, entity_name="contact")
check_contacts <- dyn_retrieve(new_contacts_result$id, 
                               entity_name="contact", columns=c("fullname"))
#clean up
deleted_contacts_result1 <- dyn_delete(updated_contacts_result1$id,
                                       entity_name="contact")
}
}
