% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{extract_key_value_data}
\alias{extract_key_value_data}
\title{Extract Key-Value Pairs}
\usage{
extract_key_value_data(x)
}
\arguments{
\item{node}{the XML node or document to be converted to an R list}
}
\value{
\code{list} parsed from the supplied node
}
\description{
This function extracts a series of key-value pairs into a single row tibble
where each key is a column and the first row contains the value for each.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
