% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{gen_value_node}
\alias{gen_value_node}
\title{Create XML Node for Single Value}
\usage{
gen_value_node(x)
}
\arguments{
\item{x}{value; a single column/attribute value from a record}
}
\value{
\code{xmlNode}
}
\description{
This function takes a single value for a record and converts it to a valid XML 
node with attribute type so that Dynamics CRM will recognize the type and value.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
