% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_fetchxml}
\alias{generate_fetchxml}
\title{Generate FetchXML from Arguments}
\usage{
generate_fetchxml(entity_name, attributes, all_attributes, page_size, top)
}
\arguments{
\item{entity_name}{character; the name of one MS Dynamics CRM entity that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{attributes}{character; one or more strings indicating attributes}

\item{all_attributes}{logical; an indicator if all possible attributes should be returned 
for the entity. If \code{TRUE} this parameter will override the \code{attributes} parameter.}

\item{page_size}{numeric; a number indicating the records per page that are 
returned. This is provided for performance improvements for queries returning a 
large number of records.}

\item{top}{numeric; a number indicating the first N number of records that should 
be returned by the query. This behavior is similar to the LIMIT or TOP keyword in SQL.}
}
\description{
This function will generate a syntactically valid FetchXML for simple "SELECT" 
statements. If more complex ordering, filtering, etc. is needed, then the user 
should specify an exact FetchXML string.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
