% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{header_available}
\alias{header_available}
\title{Check header availability}
\usage{
header_available(verbose = FALSE)
}
\value{
logical
}
\description{
Check if an authentication header is available in \code{\link{rdynamicscrm}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
