% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdynamicscrm.R
\docType{package}
\name{rdynamicscrm}
\alias{rdynamicscrm}
\alias{rdynamicscrm-package}
\title{\code{rdynamicscrm} package}
\description{
An implementation of a MS Dynamics CRM client
}
\details{
An implementation of a 'MS Dynamics CRM' client. Microsoft Dynamics CRM is a 
customer relationship management (CRM) software package developed by Microsoft. 
This package is an articulation of SOAP API calls to a CRM Online and On Premise 
IFD environment.

Additional material can be found in the 
\href{https://github.com/StevenMMortimer/rdynamicscrm}{README} on GitHub
}
