% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{update_header}
\alias{update_header}
\title{Update the Header of a Call}
\usage{
update_header(doc)
}
\arguments{
\item{doc}{the XML node or document to be converted to an R list}
}
\value{
an XML document as text
}
\description{
This function swaps in a new header for a posted body. This is mainly needed 
whenever the header expires and needs updated.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
